Поменять тип id с BigInteger на Long

+++ antifraud/src/main/java/com/bank/antifraud/controller/SuspiciousCardTransferController.java
@@ -15,7 +15,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @RestController
@@ -36,14 +35,14 @@ public class SuspiciousCardTransferController {
     }
 
     @GetMapping("{id}")
-    public ResponseEntity<SuspiciousCardTransferEntity> getById(@PathVariable BigInteger id) {
+    public ResponseEntity<SuspiciousCardTransferEntity> getById(@PathVariable Long id) {
         LOGGER.info("В SuspiciousCardTransferController сработал метод getById, с id: " + id.toString());
 
         return ResponseEntity.ok(suspiciousCardTransferService.findById(id));
     }
 
     @DeleteMapping("{id}")
-    public ResponseEntity<BigInteger> deleteAudit(@PathVariable BigInteger id) {
+    public ResponseEntity<Long> deleteAudit(@PathVariable Long id) {
         LOGGER.info("В SuspiciousCardTransferController сработал метод deleteAudit, с id: " + id.toString());
 
         suspiciousCardTransferService.delete(id);
