+++ antifraud/src/main/java/com/bank/antifraud/dto/SuspiciousPhoneTransfersDto.java
@@ -1,9 +1,7 @@
 package com.bank.antifraud.dto;
 
-import java.math.BigInteger;
-
 public record SuspiciousPhoneTransfersDto
-        (BigInteger id,
+        (Long id,
          Long phoneTransferId,
          Boolean isBlocked,
          Boolean isSuspicious,
