+++ antifraud/src/main/java/com/bank/antifraud/service/SuspiciousCardTransferService.java
@@ -3,13 +3,12 @@ package com.bank.antifraud.service;
 import com.bank.antifraud.dto.SuspiciousCardTransferDto;
 import com.bank.antifraud.entity.SuspiciousCardTransferEntity;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public interface SuspiciousCardTransferService {
     public SuspiciousCardTransferEntity save(SuspiciousCardTransferDto suspiciousCardTransferDto);
-    public SuspiciousCardTransferEntity findById(BigInteger id);
+    public SuspiciousCardTransferEntity findById(Long id);
     public List<SuspiciousCardTransferEntity> findAll();
-    public void delete(BigInteger id);
+    public void delete(Long id);
     public SuspiciousCardTransferEntity update(SuspiciousCardTransferDto suspiciousCardTransferDto);
 }
